+++ src/pydiverse/pipedag/materialize/store.py
@@ -21,6 +21,7 @@ from pydiverse.pipedag.core.task import Task, TaskGetItem
 from pydiverse.pipedag.errors import (
     CacheError,
     DuplicateNameError,
+    HashingError,
     HookCheckException,
     StageError,
     StoreIncompatibleException,
@@ -138,15 +139,23 @@ class BaseTableStore(TableHookResolver, Disposable):
         """
         config_context = ConfigContext.get()
         hook = None
+        generate_lazy_query_str_success = False
+        warning_str = None
         try:
             hook = self.get_m_table_hook(table)
             query_str = hook.lazy_query_str(self, table.obj)
+            generate_lazy_query_str_success = True
         except TypeError:
+            warning_str = (
+                f"The output table {table.name} given by a "
+                f"{repr(type(table.obj))} of the lazy task {task._name} does "
+                "not provide a query string."
+            )
+        except HashingError as e:
+            warning_str = f"The output table {table.name} of the lazy task {task._name} could not be hashed: {str(e)}."
+        if not generate_lazy_query_str_success:
             self.logger.warning(
-                f"The output table {table.name} given by a"
-                f" {repr(type(table.obj))} of the lazy task {task._name} does"
-                " not provide a query string. Lazy evaluation is not"
-                " possible. Assuming that the table is not cache valid."
+                warning_str + " Lazy evaluation is not possible. Assuming that the table is not cache valid."
             )
             # Assign random query string to ensure that task is not cache valid
             query_str = uuid.uuid4().hex